kitchen sink 11ty config, plus comments.
I don't know how feasible it is to make a copypasta of
"how to start a site, with most of the plugins you probably want",
but this is my attempt.

Ideally it becomes organized in such a way that you can comment out
full hunks of adjacent code to remove a feature, and that works;
or conversely, add a single hunk of code in one contiguous text change,
and have that result in a completely introduced feature.

That goal is not quite recognized yet.
Several functions reach across multiple uses in a way that confounds it;
these are probably fixable (and some comments point the probable way).
Also problematic is that some features are interacting with layout
templates in a way that's contained in other files, far away;
figuring out a way we could package these together is a bit beyond
what I've had time to contemplate so far.
Still, I'd like to aim in that general direction.

This is a checkpoint, and fully tested; it may need some amending.