use rooted links.
This is generally my preference when doing linking within a site,
because it's the easiest to update page locations with a regexp.

Alternatively, switching these to "./" would've worked (and this
site is so simple it really makes no matter).
However, either that change or this one would've been required...
because our replaceLink feature doesn't engage unless there's at
least one slash.  Ahheh.

(I would really like 11ty to get a core feature for this link fixup
stuff.  It's kinda core to 99% of usage, imo.)